moves files from /api to src root
