+++ src/updateContent.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { convertHtmlToNotionBlocks } from "../utils/html";
+import { convertHtmlToNotionBlocks } from "./utils/html";
 
 const notionApiToken = process.env.NOTION_API_TOKEN;
